ingestion/requirements.txt
geopandas>=0.14.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
shapely>=2.0.0
pyproj>=3.0.0
fiona>=1.9.0
geoalchemy2>=0.14.0